refactor: update the profile and details page to be server component and use cookie for reduced db calls
